==> 02_Inicjowanie_bazy_danych.ipynb <==
#import bibliotek

!pip install psycopg2 binary
import psycopg2
from psycopg2 import connect

#Połączenie z bazą danych
username = 'nick'
password = 'haslo'
host = 'localhost'
database = 'airlines'
port = port

#Zdefiniowanie zmiennej con i cursor
con = psycopg2.connect(
    dbname=database,
    user=username,
    password=password,
    host=host,
    port=port
)
cursor = con.cursor()

# Sprawdzenie, czy połączenie działa
cur = con.cursor()
cur.execute('SELECT version();')
print(cur.fetchone()[0])
cur.close()

#Wczytanie pliku z katalogu sql
with open('sciezka do katalogu', 'r') as file_schema:
    schema_sql = file_schema.read()

#Rozdzielnie zawartości pliku na kwerendy
queries = schema_sql.split(';')
queries = [query.strip() + ';' for query in queries if query.strip()]
for query in queries:
    print(query)

    
    
#Zainicjowanie struktury bazy

#Aircraft
cur.execute('''
    DROP TABLE IF EXISTS public.aircraft CASCADE;
    CREATE TABLE IF NOT EXISTS public.aircraft
    (
        id serial,
        manufacture_year bigint,
        tail_num text,
        number_of_seats double precision
    );
''')
con.commit()

#Airport_weather
cur.execute('''
    DROP TABLE IF EXISTS public.airport_weather CASCADE;
    CREATE TABLE IF NOT EXISTS public.airport_weather
    (
        id serial,
        station text,
        name text,
        date text,
        awnd double precision,
        pgtm double precision,
        prcp double precision,
        snow double precision,
        snwd double precision,
        tavg double precision,
        tmax double precision,
        tmin double precision,
        wdf2 double precision,
        wdf5 double precision,
        wsf2 double precision,
        wsf5 double precision,
        wt01 double precision,
        wt02 double precision,
        wt03 double precision,
        wt04 double precision,
        wt05 double precision,
        wt06 double precision,
        wt07 double precision,
        wt08 double precision,
        wt09 double precision,
        wesd double precision,
        wt10 double precision,
        psun double precision,
        tsun double precision,
        sn32 double precision,
        sx32 double precision,
        tobs double precision,
        wt11 double precision,
        wt18 double precision
    );
''')
con.commit()

#Flight
cur.execute('''
    DROP TABLE IF EXISTS public.flight CASCADE;
    CREATE TABLE IF NOT EXISTS public.flight
    (
        id serial,
        month bigint,
        day_of_month bigint,
        day_of_week bigint,
        op_unique_carrier text,
        tail_num text,
        op_carrier_fl_num bigint,
        origin_airport_id bigint,
        dest_airport_id bigint,
        crs_dep_time bigint,
        dep_time double precision,
        dep_delay_new double precision,
        dep_time_blk text,
        crs_arr_time bigint,
        arr_time double precision,
        arr_delay_new double precision,
        arr_time_blk text,
        cancelled double precision,
        crs_elapsed_time double precision,
        actual_elapsed_time double precision,
        distance double precision,
        distance_group bigint,
        carrier_delay double precision,
        weather_delay double precision,
        nas_delay double precision,
        security_delay double precision,
        late_aircraft_delay double precision,
        year bigint
    );
''')
con.commit()

# Airport_list
cur.execute('''
    DROP TABLE IF EXISTS public.airport_list CASCADE;
    CREATE TABLE IF NOT EXISTS public.airport_list
    (
        id serial,
        origin_airport_id bigint,
        display_airport_name text,
        origin_city_name text,
        name text
    );
''')
con.commit()

# Widok aircraft
cur.execute('''
    -- BLOCKER TO FORCE USAGE if_exists in PANDAS (REMOVE IF OBSTACLE)
    CREATE OR REPLACE VIEW public.v_aircraft_blocker AS
    SELECT 1 FROM public.aircraft LIMIT 1;
''')
con.commit()

# Widok Airport_weather
cur.execute('''
    -- BLOCKER TO FORCE USAGE if_exists in PANDAS (REMOVE IF OBSTACLE)
    CREATE OR REPLACE VIEW public.v_aircraft_weather_blocker AS
    SELECT 1 FROM public.airport_weather LIMIT 1;
''')
con.commit()

#Widok Flight
cur.execute('''
    -- BLOCKER TO FORCE USAGE if_exists in PANDAS (REMOVE IF OBSTACLE)
    CREATE OR REPLACE VIEW public.v_flight_blocker AS
    SELECT 1 FROM public.flight LIMIT 1;
''')
con.commit()

# Widok Airport_list
cur.execute('''
    CREATE OR REPLACE VIEW public.v_airport_list_blocker AS
    SELECT 1 FROM public.airport_list LIMIT 1;
''')
con.commit()

    

#Sprawdzenie

def check_if_table_exists(table_name):
    msg = f"I check if the table exists {table_name}"
    print(msg)

    query = f"select 1 from {table_name}"
    # jeżeli tabela nie istnieje, ten krok zwróci wyjątek
    cursor.execute(query)
    print('OK!')

tables_to_test = [
    'aircraft',
    'airport_weather',
    'flight',
    'airport_list'
]

for table in tables_to_test:
    check_if_table_exists(table)

con.close()
